// packages/api/jw/generated/reference-index.generated.ts
export interface ReferenceIndex {

    messages: string[];

    rules: string[];

    codelists: string[];

    constraints: string[];

    concepts: string[];

    xmlElements: string[];

    aliases: Record<string, string[]>;

}